+++ src/main/java/study/ascii92der/lotto/step3n4/ResultView.java
@@ -1,4 +1,4 @@
-package study.ascii92der.lotto.step3;
+package study.ascii92der.lotto.step3n4;
 
 public class ResultView {
